refactoring scaffolding application
