+++ packages/interop-client/src/api/index.ts
@@ -1,0 +1,1 @@
+export * from "./getAgreementByPurposeId.js";
