+++ packages/interop-client/src/models/gateway.models.ts
